+++ R/get_cluster.R
@@ -9,14 +9,19 @@ get_cluster <- function(conn) {
     "observation" %in% dbListTables(conn, "observation"),
     msg = "No observations found. Did you run `import_observations()`?"
   )
-  obs <- dbGetQuery(
-    conn, "SELECT id, x, y, survey, status, cluster FROM observation"
-  )
+  "SELECT
+o.id, o.x, o.y, o.survey, o.status, o.cluster
+FROM observation AS o
+LEFT JOIN unlikely AS u ON o.survey = u.survey
+WHERE u.value IS NULL" |>
+    dbGetQuery(conn = conn) -> obs
   cluster <- dbGetQuery(conn, "
 SELECT
-  cluster, COUNT(cluster) AS n_obs, MAX(status) AS max_status,
-  AVG(x) AS centroid_x, AVG(y) AS centroid_y
-FROM observation
-GROUP BY cluster")
+  o.cluster, COUNT(o.cluster) AS n_obs, MAX(status) AS max_status,
+  AVG(o.x) AS centroid_x, AVG(o.y) AS centroid_y
+FROM observation AS o
+LEFT JOIN unlikely AS u ON o.survey = u.survey
+WHERE u.value IS NULL
+GROUP BY o.cluster")
   return(list(cluster = cluster, observations = obs))
 }